Solid foundations are absolutely essential.
Even the shittiest houses will get their owners sued to hell and back if
they hurt someone, so their foundations are rock-solid.

Even though this analogy completely breaks down when applying it to this
situation, I'm laying a bit more groundwork to prepare for proper
development. In this commit: the Player class is created.